clizzle
This is my first working python script that lets me shoot emails through gmail using Google's api for python. It grabs the astrological data I use in each entry from die.net and lunaf.com and places a header with that data at the top of the email. The script can take two args, one being the text file with the body of the email and the second being the subject.